Added building links to circles too
