+++ lib/glassfrog/accountability.rb
@@ -1,4 +1,5 @@
 require 'glassfrog/base'
+require 'glassfrog/link_factory'
 
 module Glassfrog
   #
@@ -7,6 +8,8 @@ module Glassfrog
   class Accountability < Glassfrog::Base
     # @return [String]
     attr_accessor :description
+
+    LinkFactory.register(:accountabilities, self)
   end
 end
