+++ lib/glassfrog/circle.rb
@@ -9,11 +9,13 @@ module Glassfrog
     # @return [String]
     attr_accessor :name, :short_name, :strategy
     # @return [Hash]
-    attr_accessor :links
+    attr_accessor :links, :roles
     # @return [Array<Glassfrog::Circle]
     attr_accessor :sub_circles
     PATH = '/circles'
     TYPE = :circles
+
+    LINK_TYPES = [:roles]
 
     # 
     # Sends a GET request for Circle(s) to GlassFrog.
@@ -23,7 +25,11 @@ module Glassfrog
     # @return [Array<Glassfrog::Circle>] The array of Circle(s) fetched from GlassFrog.
     def self.get(client, options)
       response = Glassfrog::REST::Get.get(client, PATH, options)
-      response[TYPE].map { |object| self.new(object) }
+      response[TYPE].map do |object|
+        circle = self.new(object)
+        LINK_TYPES.each { |type| circle.build_link_objects(response, type) }
+        circle
+      end
     end
   end
 end
