+++ lib/glassfrog/domain.rb
@@ -1,4 +1,5 @@
 require 'glassfrog/base'
+require 'glassfrog/link_factory'
 
 module Glassfrog
   #
@@ -7,5 +8,7 @@ module Glassfrog
   class Domain < Glassfrog::Base
     # @return [String]
     attr_accessor :description
+
+    LinkFactory.register(:domains, self)
   end
 end
