+++ lib/glassfrog/link_factory.rb
@@ -1,17 +1,13 @@
-require 'glassfrog/accountability'
-require 'glassfrog/domain'
-require 'glassfrog/person'
-
 module Glassfrog
   class LinkFactory
-    CONFIG = {
-      accountabilities: Accountability,
-      domains: Domain,
-      people: Person
-    }
+    @@config = {}
+
+    def self.register(key, klass)
+      @@config[key] = klass
+    end
 
     def self.build(link_type, attributes)
-      klass = CONFIG[link_type]
+      klass = @@config[link_type]
       klass.new(attributes)
     end
   end
